+++ src/server/__tests__/utils/index.test.ts
@@ -3,6 +3,7 @@ import * as utilsExports from "../../utils/index";
 import * as loggerExports from "../../utils/logger";
 import * as rateLimitExports from "../../utils/rate-limit";
 import * as schemaExports from "../../utils/schema";
+import * as sessionExports from "../../utils/session";
 
 describe("server/utils module exports", () => {
   it("should export everything from cleanup", () => {
@@ -37,12 +38,21 @@ describe("server/utils module exports", () => {
     });
   });
 
-  it("should only export members from cleanup, logger, rate-limit, and schema", () => {
+  it("should export everything from session", () => {
+    Object.keys(sessionExports).forEach((key) => {
+      expect(utilsExports[key as keyof typeof utilsExports]).toBe(
+        sessionExports[key as keyof typeof sessionExports],
+      );
+    });
+  });
+
+  it("should only export members from cleanup, logger, rate-limit, schema, and session", () => {
     const expectedExports = [
       ...Object.keys(cleanupExports),
       ...Object.keys(loggerExports),
       ...Object.keys(rateLimitExports),
       ...Object.keys(schemaExports),
+      ...Object.keys(sessionExports),
     ];
 
     const actualExports = Object.keys(utilsExports);