+++ src/server/utils/index.ts
@@ -7,3 +7,4 @@ export * from "./cleanup";
 export * from "./logger";
 export * from "./rate-limit";
 export * from "./schema";
+export * from "./session";